address code review from Aitor
