+++ app/src/main/java/com/duckduckgo/app/browser/BrowserTabViewModel.kt
@@ -98,6 +98,7 @@ import com.duckduckgo.app.location.GeoLocationPermissions
 import com.duckduckgo.app.location.data.LocationPermissionType
 import com.duckduckgo.app.location.data.LocationPermissionsRepository
 import com.duckduckgo.app.pixels.AppPixelName
+import com.duckduckgo.app.pixels.remoteconfig.AndroidBrowserConfigFeature
 import com.duckduckgo.app.privacy.db.NetworkLeaderboardDao
 import com.duckduckgo.app.privacy.db.UserAllowListRepository
 import com.duckduckgo.app.settings.db.SettingsDataStore
@@ -214,6 +215,7 @@ class BrowserTabViewModel @Inject constructor(
     private val sitePermissionsManager: SitePermissionsManager,
     private val syncEngine: SyncEngine,
     private val cameraHardwareChecker: CameraHardwareChecker,
+    private val androidBrowserConfig: AndroidBrowserConfigFeature,
 ) : WebViewClientListener,
     EditSavedSiteListener,
     DeleteBookmarkListener,
@@ -3104,13 +3106,21 @@ class BrowserTabViewModel @Inject constructor(
 
     private fun screenLock(featureName: String, method: String, id: String, data: JSONObject) {
         viewModelScope.launch(dispatchers.main()) {
-            command.value = ScreenLock(JsCallbackData(data, featureName, method, id))
+            if (androidBrowserConfig.screenLock().isEnabled()) {
+                withContext(dispatchers.main()) {
+                    command.value = ScreenLock(JsCallbackData(data, featureName, method, id))
+                }
+            }
         }
     }
 
     private fun screenUnlock() {
         viewModelScope.launch(dispatchers.main()) {
-            command.value = ScreenUnlock
+            if (androidBrowserConfig.screenLock().isEnabled()) {
+                withContext(dispatchers.main()) {
+                    command.value = ScreenUnlock
+                }
+            }
         }
     }
